+++ test/migrators/namespace_dart_test.dart
@@ -1,0 +1,11 @@
+// Copyright 2021 Google LLC
+//
+// Use of this source code is governed by an MIT-style
+// license that can be found in the LICENSE file or at
+// https://opensource.org/licenses/MIT.
+
+import '../utils.dart';
+
+main() {
+  testMigrator("namespace");
+}