+++ test/migrators/namespace_node_test.dart
@@ -1,0 +1,16 @@
+// Copyright 2021 Google LLC
+//
+// Use of this source code is governed by an MIT-style
+// license that can be found in the LICENSE file or at
+// https://opensource.org/licenses/MIT.
+
+@Tags(["node"])
+
+import 'package:test/test.dart';
+
+import '../utils.dart';
+
+main() {
+  runNodeTests = true;
+  testMigrator("division");
+}